refactor(gerichtskosten): move gerichtskosten.ts to shared folder
