+++ app/flows/geldEinklagenVorabcheck/index.ts
@@ -1,7 +1,7 @@
 import config from "./flow.json";
 import { guards } from "./guards";
 import { type AllContexts } from "../common";
-import { getGerichtskostenvorschuss } from "../gerichtskosten";
+import { getGerichtskostenvorschuss } from "../shared/gerichtskosten";
 
 export const geldEinklagenVorabcheck = {
   cmsSlug: "vorab-check-pages",
